viewer: refactor and cleanup
