+++ tools/startVanillaServer.js
@@ -104,7 +104,7 @@ async function startServerAndWait (version, withTimeout, options) {
 
 if (!module.parent) {
   // if (process.argv.length < 3) throw Error('Missing version argument')
-  startServer(process.argv[2] || '1.16.201')
+  startServer(process.argv[2] || '1.16.201', null, process.argv[3] ? { 'server-port': process.argv[3] } : undefined)
 }
 
 module.exports = { fetchLatestStable, startServer, startServerAndWait }
